Eliminada oferta, agregada tabla evento/ocasional.

+++ TuristiandoAPP2Comfama/src/main/java/org/example/validaciones/OcasionalValidacion.java
@@ -7,7 +7,7 @@ public class OcasionalValidacion {
     public OcasionalValidacion() {
     }
 
-    public boolean validarValorCuota(Integer valorCuota) throws Exception{
+    public boolean validarValorCuota(Double valorCuota) throws Exception{
         if (valorCuota < 0){
             throw new Exception(Mensaje.COSTO_PERSONA_INVALIDO.getMensaje());
         } else if (valorCuota > 200000) {
